+++ iam.tf
@@ -4,7 +4,7 @@ resource "google_service_account" "gke_sa" {
 }
 
 # https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/google_project_iam#google_project_iam_member
-resource "google_project_iam_member" "gke_sa_iam_member" {
+resource "google_project_iam_member" "gke_sa" {
   project = var.project_id
   count   = length(var.iam_roles_list)
   role    = var.iam_roles_list[count.index]
@@ -20,7 +20,7 @@ resource "google_service_account" "wi_gsa" {
   display_name = "Workload Identity Google service account"
 }
 
-resource "google_project_iam_member" "wi_gsa_iam_member" {
+resource "google_project_iam_member" "wi_gsa" {
   project = var.project_id
   count   = length(var.wi_iam_roles_list)
   role    = var.wi_iam_roles_list[count.index]