+++ iap_proxy.tf
@@ -1,10 +1,10 @@
 # the resources here are only created
 # if enable_private_endpoint = "true"
-resource "google_compute_subnetwork" "iap_subnet" {
+resource "google_compute_subnetwork" "iap" {
   count                    = var.enable_private_endpoint ? 1 : 0
   name                     = "${var.gke_cluster_name}-iap-subnet"
   ip_cidr_range            = var.iap_proxy_ip_cidr
-  network                  = google_compute_network.k8s_vpc.id
+  network                  = google_compute_network.k8s.id
   private_ip_google_access = "true"
   region                   = var.region
 }
@@ -12,7 +12,7 @@ resource "google_compute_subnetwork" "iap_subnet" {
 resource "google_compute_firewall" "iap_tcp_forwarding" {
   count   = var.enable_private_endpoint ? 1 : 0
   name    = "allow-ingress-from-iap"
-  network = google_compute_network.k8s_vpc.name
+  network = google_compute_network.k8s.name
 
   direction = "INGRESS"
 
@@ -41,16 +41,16 @@ resource "google_compute_instance" "iap-proxy" {
     }
   }
 
-  # because we're setting a count on the iap_subnet,
+  # because we're setting a count on the iap subnet,
   # we now have to reference it with an index as well
   network_interface {
-    network    = google_compute_network.k8s_vpc.id
-    subnetwork = google_compute_subnetwork.iap_subnet[count.index].name
+    network    = google_compute_network.k8s.id
+    subnetwork = google_compute_subnetwork.iap[count.index].name
   }
 
   metadata_startup_script = file("./scripts/startup.sh")
 
   depends_on = [
-    google_compute_router_nat.k8s_vpc_router_nat
+    google_compute_router_nat.k8s_vpc
   ]
 }