+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [0.16.3] - 2024-04-02
+### Added
+- Added `description` to output values
+### Changed
+- Added `roles/artifactregistry.reader" to GKE service account IAM roles list
+- Renamed resources
+
 ## [0.16.2] - 2024-02-20
 ### Added
 - `examples/kyverno`